requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torchdiffeq>=0.2.3

# Data processing
numpy>=1.24.0
h5py>=3.8.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging and visualization
tensorboard>=2.12.0
matplotlib>=3.7.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
hypothesis>=6.75.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
